requirements.txt
Django==1.11.2
django-adminplus==0.5
django-cors==0.1
djangorestframework==3.6.3
pytz==2017.2
